Improve consumable creation workflow
This adds proper status from sidekiq to the frontend. Rather than waiting an arbitrary amount of time, we actually wait until sidekiq is done processing the tag creation job. Overall this should be done a different and less janky way. Until I have time to rebuild the asset tag creation system from the ground up, this is a good solution.